replace obejctMetadataMapsService with method in workspaceCacheStorageService
