+++ packages/twenty-server/src/engine/core-modules/search/search.module.ts
@@ -3,10 +3,10 @@ import { Module } from '@nestjs/common';
 import { FileModule } from 'src/engine/core-modules/file/file.module';
 import { SearchResolver } from 'src/engine/core-modules/search/search.resolver';
 import { SearchService } from 'src/engine/core-modules/search/services/search.service';
-import { ObjectMetadataModule } from 'src/engine/metadata-modules/object-metadata/object-metadata.module';
+import { WorkspaceCacheStorageModule } from 'src/engine/workspace-cache-storage/workspace-cache-storage.module';
 
 @Module({
-  imports: [FileModule, ObjectMetadataModule],
+  imports: [FileModule, WorkspaceCacheStorageModule],
   providers: [SearchResolver, SearchService],
 })
 export class SearchModule {}
