+++ packages/twenty-server/src/engine/workspace-cache-storage/workspace-cache-storage.service.ts
@@ -11,6 +11,14 @@ import { InjectCacheStorage } from 'src/engine/core-modules/cache-storage/decora
 import { CacheStorageService } from 'src/engine/core-modules/cache-storage/services/cache-storage.service';
 import { CacheStorageNamespace } from 'src/engine/core-modules/cache-storage/types/cache-storage-namespace.enum';
 import { ObjectMetadataMaps } from 'src/engine/metadata-modules/types/object-metadata-maps';
+import {
+  WorkspaceMetadataCacheException,
+  WorkspaceMetadataCacheExceptionCode,
+} from 'src/engine/metadata-modules/workspace-metadata-cache/exceptions/workspace-metadata-cache.exception';
+import {
+  WorkspaceMetadataVersionException,
+  WorkspaceMetadataVersionExceptionCode,
+} from 'src/engine/metadata-modules/workspace-metadata-version/exceptions/workspace-metadata-version.exception';
 
 export enum WorkspaceCacheKeys {
   GraphQLTypeDefs = 'graphql:type-defs',
@@ -131,6 +139,31 @@ export class WorkspaceCacheStorageService {
     return this.cacheStorageService.get<ObjectMetadataMaps>(
       `${WorkspaceCacheKeys.MetadataObjectMetadataMaps}:${workspaceId}:${metadataVersion}`,
     );
+  }
+
+  async getObjectMetadataMapsOrThrow(workspaceId: string) {
+    const currentCacheVersion = await this.getMetadataVersion(workspaceId);
+
+    if (currentCacheVersion === undefined) {
+      throw new WorkspaceMetadataVersionException(
+        `Metadata version not found for workspace ${workspaceId}`,
+        WorkspaceMetadataVersionExceptionCode.METADATA_VERSION_NOT_FOUND,
+      );
+    }
+
+    const objectMetadataMaps = await this.getObjectMetadataMaps(
+      workspaceId,
+      currentCacheVersion,
+    );
+
+    if (!objectMetadataMaps) {
+      throw new WorkspaceMetadataCacheException(
+        `Object metadata map not found for workspace ${workspaceId} and metadata version ${currentCacheVersion}`,
+        WorkspaceMetadataCacheExceptionCode.OBJECT_METADATA_MAP_NOT_FOUND,
+      );
+    }
+
+    return objectMetadataMaps;
   }
 
   setGraphQLTypeDefs(
